+++ Groundnut/core/src/Groundnut/core/physics/testWorld.java
@@ -1,0 +1,13 @@
+package Groundnut.core.physics;
+
+import com.badlogic.gdx.physics.box2d.Box2DDebugRenderer;
+
+public class testWorld {
+
+
+    Box2DDebugRenderer renderer = new Box2DDebugRenderer();
+
+    public testWorld() {
+
+    }
+}
