+++ box2Dtest/desktop/src/com/mygdx/game/desktop/DesktopLauncher.java
@@ -2,16 +2,11 @@ package com.mygdx.game.desktop;
 
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
-import com.mygdx.game.*;
+import com.mygdx.game.MyGdxGame;
 
 public class DesktopLauncher {
 	public static void main (String[] arg) {
 		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
-		new LwjglApplication(new GraphicsLoop(), config);
-
-		ClientListenerThread listener = new ClientListenerThread();
-		ClientWriterThread writer = new ClientWriterThread();
-		listener.start();
-		writer.start();
+		new LwjglApplication(new MyGdxGame(), config);
 	}
 }
