+++ box2Dtest/settings.gradle
@@ -1,0 +1,1 @@
+include 'desktop', 'core'
